core: refactor and border radius
