+++ components/sections/ContainerForm.tsx
@@ -99,7 +99,7 @@ const ContainerForm: FC<ContainerFormType> = (props: ContainerFormType) => {
           "w-3/12": width === 'w_3_12',
           "w-2/12": width === 'w_2_12',
         })}>
-          <section className={cn("mobile:p-3 p-6 shadow-15 bg-surface-0 relative", { "shadow-none": shadow == false })}>
+          <section className={cn("mobile:p-3 p-6 shadow-15 bg-surface-0 relative rounded-lg", { "shadow-none": shadow == false })}>
             {
               isLoading
                 ? <div className="absolute w-full h-full z-20 flex justify-center items-center left-0 top-0 bg-surface-0">
